Refactor registration routes to a different file
